+++ index.html
@@ -229,6 +229,6 @@
     </div>
   </template>
 
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
