+++ js/constants.js
@@ -1,0 +1,17 @@
+const MIN_PRICE = 1000; //Минимальная цена
+const MAX_PRICE = 100000; //Максимальная цена
+const LAT_MIN = 35.65000; //Координаты. Минимальная широта
+const LAT_MAX = 35.70000; //Координаты. Максимальная широта
+const LNG_MIN = 139.70000; //Координаты. Минимальная долгота
+const LNG_MAX = 139.80000; //Координаты. Максимальная долгота
+const ARRAY_LENGTH = 10; //Длина массива с объектами
+
+export {
+  MIN_PRICE,
+  MAX_PRICE,
+  LAT_MIN,
+  LAT_MAX,
+  LNG_MIN,
+  LNG_MAX,
+  ARRAY_LENGTH,
+};
