+++ js/variables.js
@@ -1,0 +1,16 @@
+const types = ['palace', 'flat', 'house', 'bungalow', 'hotel']; //Тип жилья
+const checkinTimes = ['12:00', '13:00', '14:00']; //Время заезда
+const checkoutTimes = ['12:00', '13:00', '14:00']; //Время выезда
+const features = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner']; //Доп. опции
+const photos = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg', //Фотографии
+];
+
+export {
+  types,
+  checkinTimes,
+  checkoutTimes,
+  features,
+  photos
+};
